exfat: move helpers functions to misc.c
