+++ fs/exfat/exfat_fs.h
@@ -469,10 +469,6 @@ int exfat_get_cluster(struct inode *inode, unsigned int cluster,
 /* dir.c */
 extern const struct inode_operations exfat_dir_inode_operations;
 extern const struct file_operations exfat_dir_operations;
-void exfat_update_bh(struct super_block *sb, struct buffer_head *bh, int sync);
-void exfat_chain_set(struct exfat_chain *ec, unsigned int dir,
-		unsigned int size, unsigned char flags);
-struct exfat_chain *exfat_chain_dup(struct exfat_chain *dir);
 void exfat_get_uniname_from_ext_entry(struct super_block *sb,
 		struct exfat_chain *p_dir, int entry, unsigned short *uniname);
 unsigned int exfat_get_entry_type(struct exfat_dentry *p_entry);
@@ -541,5 +537,9 @@ struct exfat_timestamp *exfat_tm_now(struct exfat_sb_info *sbi,
 		struct exfat_timestamp *tm);
 unsigned short exfat_calc_chksum_2byte(void *data, int len,
 		unsigned short chksum, int type);
+void exfat_update_bh(struct super_block *sb, struct buffer_head *bh, int sync);
+void exfat_chain_set(struct exfat_chain *ec, unsigned int dir,
+		unsigned int size, unsigned char flags);
+struct exfat_chain *exfat_chain_dup(struct exfat_chain *dir);
 
 #endif /* !_EXFAT_H */
